Mergerrig.....
(warning this is unstable code so maybe some function will broken)


lint

+++ bot/helper/mirror_utils/download_utils/qbit_downloader.py
@@ -5,10 +5,10 @@
 import logging
 import os
 import random
+import re
 import shutil
 import string
 import time
-from urllib.parse import parse_qs, urlparse
 
 import qbittorrentapi as qba
 from telegram import InlineKeyboardMarkup
@@ -35,45 +35,50 @@ from bot.helper.ext_utils.bot_utils import (
 from bot.helper.mirror_utils.status_utils.qbit_download_status import QbDownloadStatus
 from bot.helper.mirror_utils.upload_utils.gdriveTools import GoogleDriveHelper
 from bot.helper.telegram_helper import button_build
-from bot.helper.telegram_helper.message_utils import *
+from bot.helper.telegram_helper.message_utils import (
+    deleteMessage,
+    sendMarkup,
+    sendMessage,
+    sendStatusMessage,
+)
 
 LOGGER = logging.getLogger(__name__)
-logging.getLogger('qbittorrentapi').setLevel(logging.ERROR)
-logging.getLogger('requests').setLevel(logging.ERROR)
-logging.getLogger('urllib3').setLevel(logging.ERROR)
+logging.getLogger("qbittorrentapi").setLevel(logging.ERROR)
+logging.getLogger("requests").setLevel(logging.ERROR)
+logging.getLogger("urllib3").setLevel(logging.ERROR)
 
 
 class QbitTorrent:
     def __init__(self):
-        self.update_interval = 2
+        self.update_interval = 4
+        self.client = get_client()
         self.meta_time = time.time()
         self.stalled_time = time.time()
-        self.sizechecked = False
-        self.dupchecked = False
+        self.sizeChecked = False
+        self.dupChecked = False
+        self.is_file = False
+        self.pincode = ""
+        self.get_info = 0
 
     @new_thread
     def add_torrent(self, link, dire, listener, qbitsel):
-        self.client = get_client()
         self.listener = listener
         self.dire = dire
         self.qbitsel = qbitsel
-        is_file = False
-        count = 0
-        pincode = ""
         try:
             if os.path.exists(link):
-                is_file = True
+                self.is_file = True
                 self.ext_hash = get_hash_file(link)
             else:
                 self.ext_hash = get_hash_magnet(link)
             tor_info = self.client.torrents_info(torrent_hashes=self.ext_hash)
             if len(tor_info) > 0:
                 sendMessage(
-                    "Torrent ini sudah ada dalam daftar.", listener.bot, listener.update
+                    "This Torrent is already in list.", listener.bot, listener.update
                 )
                 self.client.auth_log_out()
                 return
-            if is_file:
+            if self.is_file:
                 op = self.client.torrents_add(torrent_files=[link], save_path=dire)
                 os.remove(link)
             else:
@@ -85,7 +90,7 @@ class QbitTorrent:
                     while True:
                         if time.time() - self.meta_time >= 20:
                             sendMessage(
-                                "Torrent tidak ditambahkan. Laporkan saat Anda melihat kesalahan ini",
+                                "The Torrent was not added. Report when you see this error",
                                 listener.bot,
                                 listener.update,
                             )
@@ -101,7 +106,7 @@ class QbitTorrent:
                             break
             else:
                 sendMessage(
-                    "Ini adalah tautan yang tidak didukung/tidak valid.",
+                    "This is an unsupported/invalid link.",
                     listener.bot,
                     listener.update,
                 )
@@ -112,7 +117,7 @@ class QbitTorrent:
                 return
             tor_info = tor_info[0]
             self.ext_hash = tor_info.hash
-            gid = ''.join(
+            gid = "".join(
                 random.SystemRandom().choices(
                     string.ascii_letters + string.digits, k=14
                 )
@@ -121,12 +126,12 @@ class QbitTorrent:
                 download_dict[listener.uid] = QbDownloadStatus(
                     gid, listener, self.ext_hash, self.client
                 )
-            LOGGER.info(f"QbitDownload started: {tor_info.name}")
+            LOGGER.info(f"QbitDownload started: {tor_info.name} {self.ext_hash}")
             self.updater = setInterval(self.update_interval, self.update)
             if BASE_URL is not None and qbitsel:
-                if not is_file:
+                if not self.is_file:
                     meta = sendMessage(
-                        "Mengunduh Metadata... Harap tunggu lalu Anda dapat memilih file atau mencerminkan file Torrent jika memiliki seeder rendah",
+                        "Downloading Metadata, wait then you can select files or mirror Torrent file if it have low seeders",
                         listener.bot,
                         listener.update,
                     )
@@ -150,45 +155,42 @@ class QbitTorrent:
                         except:
                             deleteMessage(listener.bot, meta)
                             return False
-                time.sleep(0.5)
                 self.client.torrents_pause(torrent_hashes=self.ext_hash)
                 for n in str(self.ext_hash):
                     if n.isdigit():
-                        pincode += str(n)
-                        count += 1
-                    if count == 4:
+                        self.pincode += str(n)
+                    if len(self.pincode) == 4:
                         break
-                URL = f"{BASE_URL}/slam/files/{self.ext_hash}"
-                pindata = f"pin {gid} {pincode}"
-                donedata = f"done {gid} {self.ext_hash}"
                 buttons = button_build.ButtonMaker()
-                buttons.buildbutton("Pilih File", URL)
-                buttons.sbutton("Kode Pin", pindata)
-                buttons.sbutton("Selesai Memilih", donedata)
+                buttons.buildbutton(
+                    "Select Files", f"{BASE_URL}/app/files/{self.ext_hash}"
+                )
+                buttons.sbutton("Pincode", f"pin {gid} {self.pincode}")
+                buttons.sbutton("Done Selecting", f"done {gid} {self.ext_hash}")
                 QBBUTTONS = InlineKeyboardMarkup(buttons.build_menu(2))
-                msg = "Unduhan Anda dijeda. Pilih file lalu tekan tombol Selesai Memilih untuk mulai mengunduh."
+                msg = "Your download paused. Choose files then press Done Selecting button to start downloading."
                 sendMarkup(msg, listener.bot, listener.update, QBBUTTONS)
             else:
                 sendStatusMessage(listener.update, listener.bot)
         except qba.UnsupportedMediaType415Error as e:
             LOGGER.error(str(e))
             sendMessage(
-                "Ini adalah tautan yang tidak didukung/tidak valid: {str(e)}",
+                "This is an unsupported/invalid link: {str(e)}",
                 listener.bot,
                 listener.update,
             )
-            self.client.torrents_delete(torrent_hashes=self.ext_hash, delete_files=True)
             self.client.auth_log_out()
         except Exception as e:
-            LOGGER.error(str(e))
             sendMessage(str(e), listener.bot, listener.update)
             self.client.auth_log_out()
 
     def update(self):
         tor_info = self.client.torrents_info(torrent_hashes=self.ext_hash)
         if len(tor_info) == 0:
-            self.client.auth_log_out()
-            self.updater.cancel()
+            self.get_info += 1
+            if self.get_info > 10:
+                self.client.auth_log_out()
+                self.updater.cancel()
             return
         try:
             tor_info = tor_info[0]
@@ -208,7 +210,7 @@ class QbitTorrent:
                 if (
                     STOP_DUPLICATE
                     and not self.listener.isLeech
-                    and not self.dupchecked
+                    and not self.dupChecked
                     and os.path.isdir(f"{self.dire}")
                 ):
                     LOGGER.info("Checking File/Folder if already in Drive")
@@ -221,12 +223,12 @@ class QbitTorrent:
                         gd = GoogleDriveHelper()
                         qbmsg, button = gd.drive_list(qbname, True)
                         if qbmsg:
-                            msg = "File/Folder sudah tersedia di Drive."
+                            msg = "File/Folder is already available in Drive."
                             self.client.torrents_pause(torrent_hashes=self.ext_hash)
                             time.sleep(0.3)
                             self.listener.onDownloadError(msg)
                             sendMarkup(
-                                "Berikut adalah hasil pencariannya:",
+                                "Here are the search results:",
                                 self.listener.bot,
                                 self.listener.update,
                                 button,
@@ -235,31 +237,31 @@ class QbitTorrent:
                             self.client.auth_log_out()
                             self.updater.cancel()
                             return
-                    self.dupchecked = True
-                if not self.sizechecked:
+                    self.dupChecked = True
+                if not self.sizeChecked:
                     limit = None
                     if ZIP_UNZIP_LIMIT is not None and (
                         self.listener.isZip or self.listener.extract
                     ):
-                        mssg = f"Batas zip/unzip adalah {ZIP_UNZIP_LIMIT}GB"
+                        mssg = f"Zip/Unzip limit is {ZIP_UNZIP_LIMIT}GB"
                         limit = ZIP_UNZIP_LIMIT
                     elif TORRENT_DIRECT_LIMIT is not None:
-                        mssg = f"Batas Torrent adalah {TORRENT_DIRECT_LIMIT}GB"
+                        mssg = f"Torrent limit is {TORRENT_DIRECT_LIMIT}GB"
                         limit = TORRENT_DIRECT_LIMIT
                     if limit is not None:
                         LOGGER.info("Checking File/Folder Size...")
                         time.sleep(1)
                         size = tor_info.size
-                        if size > limit * 1024 ** 3:
+                        if size > limit * 1024**3:
                             self.client.torrents_pause(torrent_hashes=self.ext_hash)
                             time.sleep(0.3)
                             self.listener.onDownloadError(
-                                f"{mssg}.\nUkuran File/Folder Anda adalah {get_readable_file_size(size)}"
+                                f"{mssg}.\nYour File/Folder size is {get_readable_file_size(size)}"
                             )
                             self.client.torrents_delete(torrent_hashes=self.ext_hash)
                             self.client.auth_log_out()
                             self.updater.cancel()
-                    self.sizechecked = True
+                    self.sizeChecked = True
             elif tor_info.state == "stalledDL":
                 if (
                     time.time() - self.stalled_time >= 999999999
@@ -274,19 +276,25 @@ class QbitTorrent:
                 self.client.torrents_pause(torrent_hashes=self.ext_hash)
                 time.sleep(0.3)
                 self.listener.onDownloadError(
-                    "Tidak ada cukup ruang untuk torrent ini di perangkat"
+                    "No enough space for this torrent on device"
                 )
                 self.client.torrents_delete(torrent_hashes=self.ext_hash)
                 self.client.auth_log_out()
                 self.updater.cancel()
-            elif tor_info.state == "uploading" or tor_info.state.lower().endswith("up"):
+            elif tor_info.state != "checkingUP" and (
+                tor_info.state == "uploading" or tor_info.state.lower().endswith("up")
+            ):
                 self.client.torrents_pause(torrent_hashes=self.ext_hash)
                 if self.qbitsel:
                     for dirpath, subdir, files in os.walk(
                         f"{self.dire}", topdown=False
                     ):
                         for filee in files:
-                            if filee.endswith(".!qB"):
+                            if (
+                                filee.endswith(".!qB")
+                                or filee.endswith(".parts")
+                                and filee.startswith(".")
+                            ):
                                 os.remove(os.path.join(dirpath, filee))
                         for folder in subdir:
                             if folder == ".unwanted":
@@ -300,9 +308,11 @@ class QbitTorrent:
                 self.client.torrents_delete(torrent_hashes=self.ext_hash)
                 self.client.auth_log_out()
                 self.updater.cancel()
-        except:
-            self.client.auth_log_out()
-            self.updater.cancel()
+        except (IndexError, NameError):
+            self.get_info += 1
+            if self.get_info > 10:
+                self.client.auth_log_out()
+                self.updater.cancel()
 
 
 def get_confirm(update, context):
@@ -310,36 +320,28 @@ def get_confirm(update, context):
     user_id = query.from_user.id
     data = query.data
     data = data.split(" ")
-    qdl = getDownloadByGid(data[1])
-    if qdl is None:
-        query.answer(text="Tugas ini telah dibatalkan!", show_alert=True)
+    qbdl = getDownloadByGid(data[1])
+    if qbdl is None:
+        query.answer(text="This task has been cancelled!", show_alert=True)
         query.message.delete()
-
-    elif user_id != qdl.listener.message.from_user.id:
-        query.answer(text="Jangan buang waktumu!", show_alert=True)
+    elif user_id != qbdl.listener.message.from_user.id:
+        query.answer(text="Don't waste your time!", show_alert=True)
     elif data[0] == "pin":
         query.answer(text=data[2], show_alert=True)
     elif data[0] == "done":
         query.answer()
-        qdl.client.torrents_resume(torrent_hashes=data[2])
-        sendStatusMessage(qdl.listener.update, qdl.listener.bot)
+        qbdl.client.torrents_resume(torrent_hashes=data[2])
+        sendStatusMessage(qbdl.listener.update, qbdl.listener.bot)
         query.message.delete()
 
 
 def get_hash_magnet(mgt):
     if mgt.startswith("magnet:"):
-        _, _, _, _, query, _ = urlparse(mgt)
-    qs = parse_qs(query)
-    v = qs.get("xt", None)
-    if v is None or v == []:
-        LOGGER.error('Invalid magnet URI: no "xt" query parameter.')
-        return
-    v = v[0]
-    if not v.startswith("urn:btih:"):
-        LOGGER.error('Invalid magnet URI: "xt" value not valid for BitTorrent.')
-        return
-    mgt = v[len("urn:btih:") :]
-    return mgt.lower()
+        try:
+            mHash = re.search(r"xt=urn:btih:(.*)&dn=", mgt).group(1)
+        except:
+            mHash = re.search(r"xt=urn:btih:(.*)", mgt).group(1)
+        return mHash.lower()
 
 
 def get_hash_file(path):
