+++ bot/helper/mirror_utils/upload_utils/pyrogramEngine.py
@@ -1,9 +1,6 @@
-# Implement By - @anasty17 (https://github.com/SlamDevs/slam-mirrorbot/commit/d888a1e7237f4633c066f7c2bbfba030b83ad616)
-# (c) https://github.com/SlamDevs/slam-mirrorbot
-# All rights reserved
-
 import logging
 import os
+import threading
 import time
 
 from pyrogram.errors import FloodWait
@@ -62,6 +59,7 @@ class TgUploader:
         self.uploaded_bytes = 0
         self.last_uploaded = 0
         self.start_time = time.time()
+        self.__resource_lock = threading.RLock()
         self.is_cancelled = False
         self.chat_id = listener.message.chat.id
         self.message_id = listener.uid
@@ -69,10 +67,10 @@ class TgUploader:
         self.as_doc = AS_DOCUMENT
         self.thumb = f"Thumbnails/{self.user_id}.jpg"
         self.sent_msg = self.__app.get_messages(self.chat_id, self.message_id)
+        self.msgs_dict = {}
+        self.corrupted = 0
 
     def upload(self):
-        msgs_dict = {}
-        corrupted = 0
         path = f"{DOWNLOAD_DIR}{self.message_id}"
         self.user_settings()
         for dirpath, subdir, files in sorted(os.walk(path)):
@@ -84,16 +82,18 @@ class TgUploader:
                 up_path = os.path.join(dirpath, filee)
                 fsize = os.path.getsize(up_path)
                 if fsize == 0:
-                    corrupted += 1
+                    self.corrupted += 1
                     continue
                 self.upload_file(up_path, filee, dirpath)
                 if self.is_cancelled:
                     return
-                msgs_dict[filee] = self.sent_msg.message_id
+                self.msgs_dict[filee] = self.sent_msg.message_id
                 self.last_uploaded = 0
                 time.sleep(1.5)
         LOGGER.info(f"Leech Done: {self.name}")
-        self.__listener.onUploadComplete(self.name, None, msgs_dict, None, corrupted)
+        self.__listener.onUploadComplete(
+            self.name, None, self.msgs_dict, None, self.corrupted
+        )
 
     def upload_file(self, up_path, filee, dirpath):
         if CUSTOM_FILENAME is not None:
@@ -114,7 +114,12 @@ class TgUploader:
                     if thumb is None:
                         thumb = take_ss(up_path)
                         if self.is_cancelled:
-                            os.remove(thumb)
+                            if (
+                                self.thumb is None
+                                and thumb is not None
+                                and os.path.lexists(thumb)
+                            ):
+                                os.remove(thumb)
                             return
                     if not filee.upper().endswith(("MKV", "MP4")):
                         filee = os.path.splitext(filee)[0] + ".mp4"
@@ -163,7 +168,12 @@ class TgUploader:
                 if filee.upper().endswith(VIDEO_SUFFIXES) and thumb is None:
                     thumb = take_ss(up_path)
                     if self.is_cancelled:
-                        os.remove(thumb)
+                        if (
+                            self.thumb is None
+                            and thumb is not None
+                            and os.path.lexists(thumb)
+                        ):
+                            os.remove(thumb)
                         return
                 self.sent_msg = self.sent_msg.reply_document(
                     document=up_path,
@@ -190,9 +200,10 @@ class TgUploader:
         if self.is_cancelled:
             self.__app.stop_transmission()
             return
-        chunk_size = current - self.last_uploaded
-        self.last_uploaded = current
-        self.uploaded_bytes += chunk_size
+        with self.__resource_lock:
+            chunk_size = current - self.last_uploaded
+            self.last_uploaded = current
+            self.uploaded_bytes += chunk_size
 
     def user_settings(self):
         if self.user_id in AS_DOC_USERS:
@@ -211,4 +222,4 @@ class TgUploader:
     def cancel_download(self):
         self.is_cancelled = True
         LOGGER.info(f"Cancelling Upload: {self.name}")
-        self.__listener.onUploadError("unggahan Anda telah dihentikan!")
+        self.__listener.onUploadError("your upload has been stopped!")
